Complete foundations exercise 06-cascade-fix
I rearranged the declaration blocks in the stylesheet to resolve issues
with the cascade as much as possible, achieving the desired outcome.
I tweaked the specificity of selectors when simply reordering the
declaration blocks was not enough. My work reasonably matches the
desired-outcome.jpg as well as the sample solution stylesheet.