Complete exercise 05-flex-modal
I used HTML, flexbox properties, and vanilla CSS to achieve a solution
reasonably similar to the desired-outcome.png I really struggled with
this one and tried a couple different approaches that ended up being a
waste of time. After using the self-check to tune up my solution, I
checked it against the provided solution file and confirmed that they
are reasonably similar.