Complete exercise 07-flex-layout-2
I used html to group items and flexbox and CSS to arrange and style
them in order to achieve the desired-outcome.png and the behavior
demonstrated by the desired-outcome-smaller.png. Using the self check,
the desired-outcome pngs, and the solution files I determined that I
achieved a reasonably similar result.